+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# 0.2.0
+
+## Features
+
+- add `group send -f` command, allow user to send message from file
+
+
+## Bug Fixes
+
+- Fix `os.mknod` can't use in win platform
+
+## Breaking Changes
+
+- `group_send_image` changed into `group_send_mediagit `
+
 # 0.1.0
 
 ## Features